Development

+++ UniversityEnrollmentSystem/Services/Students/StudentService.cs
@@ -17,12 +17,16 @@ namespace UniversityEnrollmentSystem.Services.Students
             {
                 throw new Exception("Student number must be unique");
             }
-            _studentRepository.AddStudent(student);
+            await _studentRepository.AddStudent(student);
         }
 
         public async Task DeleteStudent(int id)
         {
-            _studentRepository.DeleteStudent(id);
+            if (!await _studentRepository.ExistsAsync(id))
+            {
+                throw new Exception("Student not found");
+            }
+            await _studentRepository.DeleteStudent(id);
         }
 
         public Task<List<Student>> GetAllStudents()
@@ -37,7 +41,13 @@ namespace UniversityEnrollmentSystem.Services.Students
 
         public async Task UpdateStudent(Student student)
         {
-            _studentRepository.UpdateStudent(student);
+            var exists = await _studentRepository.ExistsAsync(student.StudentId);
+
+            if (!exists)
+            {
+                throw new Exception($"Student with ID {student.StudentId} not found.");
+            }
+            await _studentRepository.UpdateStudent(student);
         }
 
         public async Task<Student> GetStudentByStudentNumber(int studentNumber)
